phase2/backend/src/database/connection.py
from sqlmodel import create_engine
from sqlalchemy import MetaData
import os
from dotenv import load_dotenv

load_dotenv()

# Get database URL from environment variable
DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./todo_app.db")

# Create engine
engine = create_engine(DATABASE_URL, echo=True)

# Create metadata instance
metadata = MetaData()